+++ spec/misc/anresource.js
@@ -1,11 +1,18 @@
 module.exports = function( autohost ) {
-	return {
-		name: 'three',
-		actions: {
-			c: {
-			},
-			d: {
+	return [
+		{
+			name: 'three',
+			actions: {
+				c: {},
+				d: {}
+			}
+		},
+		{
+			name: 'four',
+			actions: {
+				e: {},
+				f: {}
 			}
 		}
-	};
+	];
 };